-Already earlier, redefined the Neighbor kind to Target kind.
-Changed back to addresses between -Q and 2Q-1, for the colony-bridge complex.

-In connection with this, eliminated the Bridge kind, since it can be determined
from the addresses.

-Defined as extended colony only the colony with its bridge.  The extended base
colony and target together are called most of the time the workspace anyway.

-As an attempt to deal with alignment, changed the Patching Lemma in such a
way that it extends both ends of the area to (or near to) an address divisible
by E.  This may be separate on the left and right sides.  The algorithm
in the Patching Lemma is doing this rather cautiously, I hope that it will be
sufficient for disturbed recovery.